Delete last created book

+++ src/services/bookTracker.service.ts
@@ -50,11 +50,23 @@ class BookTrackerService {
    * @param {Role} role - The user role
    * @returns A promise
    */
-  async deleteBook(bookId: number, role: Role) {
+  async deleteLastBook(bookId: number, role: Role) {
     if (role === "ADMIN") {
+      const lastBook = await prisma.bookTracker.findFirst({
+        where: {
+          bookId
+        },
+        orderBy: {
+          createdAt: 'desc'
+        }
+      });
+      if (!lastBook) {
+        throw new HttpError("No book found to delete", 404);
+      }
+  
       return await prisma.bookTracker.delete({
         where: {
-          id: bookId,
+          id: lastBook.id,
         },
       });
     } else {
